Update push script to support pushing to feature branches
The previous push command did not allow us to push to feature
branches in a convenient way.
I amended it to make that job simpler, but did it in a very rudamentary
way. I'm not planning on fixing this right now, but if people start
using this fork more, I might make it easier to use.